tinymembench: package not supported on MIPS64
Fixes:

   http://autobuild.buildroot.net/results/bab/bab68bcf8714f215ac0b0c2546fa06608377fbb0/

This package is only supported on MIPS32 platforms. The mips-32.S file
included in the source code of this package only makes sense to be compiled
for MIPS32, and also contains hardcoded MIPS32 instructions which are
illegal on MIPS64.  Trying to compile this package on a MIPS64 platform will
end with a linkage failure due to undefined references.

Signed-off-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Reviewed-by: Markos Chandras <markos.chandras@imgtec.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>