docs/manual: limit the depth of the TOC in the HTML output
The current TOC level is set to 4, which does not mean 'a 4-level deep TOC',
but really means 'a TOC deep to the 4th level', with the first level being
level 0, which means we have:

    12. Appendix
        12.1. Makedev syntax documentation
        12.2. Makeuser syntax documentation
        12.3. Partition table layout description syntax
            12.3.1. The global section
                Properties for the global section

Which a bit too much. And yet, the fifth level is not shown, since we
don't have any! :-/

Limit the depth of the TOC to three levels, which is just enough to be
usefull, yet not cluttered by low-level titles.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>