gnutls: fix libopts detection causing build failures
Following the recent bump of gnutls to version 3.2.8, the build
started to fail on some machines where libopts is installed on the
system: gnutls configure script was incorrectly assuming that libopts
was available.

Since we don't have a package in Buildroot, this commit tells gnutls
to use its builtin libopts version.

Fixes (tested on gcc20):

  http://autobuild.buildroot.org/results/18f/18f61b3be6aed73f83b449b5082492a4a6ba8ffb/build-end.log

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>