fio: prevent selection with unusable uClibc versions
The fio program uses posix_madvise(). However, posix_madvise() is only
available in the Git tree of uClibc, and in Buildroot's version of
uClibc 0.9.33, thanks to the huge number of backported patches that we
carry.

Therefore, trying to build fio with an external uClibc toolchain is
most likely going to fail (as the uClibc version is most likely an
official stable release, and no stable release of uClibc ever had the
support for posix_madvise()). And similarly, building fio with uClibc
0.9.31 and 0.9.32 is always going to fail. We disable those use cases
to avoid repeated autobuilder failures.

No kconfig comment has been added, because we don't have a
well-defined way of specifying such exceptions.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>