package: deprecate some more development tools
Since some time, we have removed the support to build a toolchain for
the target, and therefore the support for several development tools on
the target.

This commit deprecates a few additional development tools: m4, bison,
flex and gob2. For flex, we retain the ability to build libfl, we only
deprecated the ability to build the flex binary itself.

The original motivation for this patch is that m4 is causing build
issues in some configurations, but there isn't really much incentive
to fix this package for the target, since it is not really useful for
embedded Linux systems.

Bison, Flex and Gob2 are deprecated because they are reverse
dependencies of m4.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: "Yann E. MORIN" <yann.morin.1998@free.fr>