vlc: add warning that it has known issues
We've seen autobuilder failures for vlc on many platforms and we don't
know how to solve them. It is also currently orphaned, i.e. there is
nobody who is actively using the package and maintaining it in
buildroot.

We could mark the package as BROKEN, but that would make it completely
invisible for the users and the autobuilders, which would not help its
orphan-ness. Therefore, instead, add a big fat warning in the config
menus that it has known issues.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>