libubox: fix build with lua option
Fixes
http://autobuild.buildroot.org/results/fc6/fc64896ca5c3b05dfe4b5c5e7d25fc4b6c9976ad/
and a bunch of similar failures.

When Lua package is selected, libubox sets an option to build with Lua
binding. Unfortunately, if LUAPATH is not passed libubox seeks it on
the host instead of staging due to hard coded pkg-config commands in
its CMake file. As a result Lua is not found and the build fails.
Fix it by passing LUAPATH in LIBUBOX_CONF_OPT.

Similarly, libubox seeks LUA_CFLAGS with the host's pkg-config for the
same reason. To prevent potential errors, fix it by passing LUA_CFLAGS
in LIBUBOX_CONF_OPT.

Finally, libubox fails to build with Lua 5.2 because it uses functions
removed from this version.
Fix it by activating the option only with Lua 5.1.

Signed-off-by: Hadrien Boutteville <hadrien.boutteville@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>