package/libdrm: radeon does not depend on x86
Thomas reported that the libdrm-radeon dependency on x86 (32- or 64-bit)
dates back to the time where it also depended on libpciaccess.

Now that the libpciacess dependency has been dropped, Thomas managed to
build a libdrm-radeon on something else than an x86 (an ARM for example).

Make libdrm-radeon no longer depend on x86

As a side effect, this fixes some potential "unmet direct dependencies"
from the xdriver_xf86-video-ati package (which was the initial problem
I was trying to solve.)

Reported-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>