flann: disallow static library build
Even though flann build system uses CMake, it always tries to build a
shared library, even if BUILD_SHARED_LIBS=OFF is passed. It could
probably be fixed in flann's CMakeLists.txt, but that's not really
worth the effort if upstream doesn't support it.

Therefore, we simply disallow the flann package in purely static
builds.

Fixes:

  http://autobuild.buildroot.org/results/a4e/a4ec0e9f28cd12ce770718fb407bbb4dc93b528b/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>