gdb: switch to 7.x for Blackfin
For some reason, we were keeping gdb version 6.6a specially for
Blackfin. However, it turns out that support for Blackfin was merged
in gdb 7.4 (cross-gdb and gdbserver, not native gdb on the
target). Therefore, we can simply remove the support for version 6.6a
and use 7.5 as the default.

Both 7.4 and 7.5 were built tested, including cross-gdb and gdbserver.

The original reason to switch to 7.x is that 6.6a doesn't build for
Blackfin FLAT, and while it builds for Blackfin FDPIC, it only builds
libiberty.a and does not actually build a cross debugger.

Fixes:

  http://autobuild.buildroot.org/results/b47/b47f85553336b7f63ee4ecdf8598374ce4a225a3/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>